# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.HTTPStatus:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  dto.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
      - email
      - name
      - password
    type: object
  dto.UserRegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
host: prod-team-22-t62v97db.final.prodcontest.ru
info:
  contact: { }
  description: API of LoyalT - platform for creating and managing loyalty programs
  termsOfService: http://swagger.io/terms/
  title: LoyalT API
  version: "1.0"
paths:
...
  /user/auth/register:
    post:
      consumes:
        - application/json
      description: Register a new user in the system
      parameters:
        - description: User registration data
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      summary: Register new user
      tags:
        - Auth
...
schemes:
  - https
swagger: "2.0"